Make killing configurable in startup.sh
Former-commit-id: 6a8daf166f27e00e2cf1a6864212d7bd94be5a73 [formerly 7d6093b055c64645a060608e0af0889120dd7e44] [formerly 2fa0a90596ba6c68d2c7ab9a9f115446190cb04b [formerly 8839f0771a83dfe9899896c65777887c2ce22dcc]]
Former-commit-id: 3c907b720aba146369bce864be9214cf6c64c1a7 [formerly 69cf18e1e9c3a94fda6b132026b614c5cab6e1ea]
Former-commit-id: abac6beda17b91607df5643b8b4b8ef0442fcaff

+++ bin/startup.sh
@@ -8,11 +8,13 @@ logfile="$logdir/startup.log"
 start_process() {
     local procname="$1"
     local procargs="$2"
+    local kill="$3"
 
     if [[ ! "$(which $procname)" = "" ]]; then
-        killall -q "$procname"
-
-        while pgrep -x "$procname" >/dev/null; do sleep 0.1; done
+        if [[ "$kill" = "1" ]]; then
+            killall -q "$procname"
+            while pgrep -x "$procname" >/dev/null; do sleep 0.1; done
+        fi
 
         $procname $procargs >> "$logfile" 2>&1 &
     else
@@ -25,7 +27,7 @@ start_process() {
 }
 
 echo "Starting up user processes $(date)" >> "$logfile" 2>&1
-start_process "polybar" "top"
-start_process "insync" "start"
-start_process "wpr" ""
+start_process "polybar" "top" "1"
+start_process "insync" "start" "1"
+start_process "wpr" "" "1"
